# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salus API",
        "description": "API documentation for the Salus health assistant application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Application server"
        }
    ],
    "paths": {
...
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "List the authenticated user's appointments",
                "operationId": "listAppointments",
                "responses": {
                    "200": {
                        "description": "Appointments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "doctor_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "appointment_date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-03-30T14:30:00.000000Z"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "confirmed"
                                                    },
                                                    "notes": {
                                                        "type": "string",
                                                        "example": "Follow-up consultation."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "operationId": "storeAppointment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "doctor_id",
                                    "appointment_date"
                                ],
                                "properties": {
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-03-30 14:30:00"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "First consultation for recurring headaches."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "doctor_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "appointment_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-03-30T14:30:00.000000Z"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "example": "First consultation for recurring headaches."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Sanctum bearer token. Example: Bearer 1|abcdef...",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}